started refactoring and files renaming
