+++ inventory.py
